requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6

# OpenAI (LLM + Whisper)
openai==1.65.2

# Image processing
Pillow==11.1.0

# Frontend
streamlit==1.43.1
audio-recorder-streamlit==0.0.10
requests==2.32.3
